improve the system prompt
The hardest thing I'm having to deal with here is the asterisk stage
direction speaking voice, which I understand is common from lots of
internet examples of roleplaying. It will be hard to suppress that.
Perhaps I will have to get more forceful with the instructions,
because the LLMs are sometimes still doing it even after this prompt
improvement.